backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pillow==10.4.0
torch>=2.0.0
transformers==4.42.3
opencv-python-headless==4.10.0.84
python-multipart==0.0.9
numpy>=1.26.0
pydantic==2.8.2
pydantic-settings==2.4.0
